% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_eload_temp_df_to_hourly.R
\name{agg_eload_temp_df_to_hourly}
\alias{agg_eload_temp_df_to_hourly}
\title{Aggregate 15 minute dataframe to hourly interval data}
\usage{
agg_eload_temp_df_to_hourly(eload_data, temp_data, data_is_quantity = T)
}
\arguments{
\item{eload_data}{A tidy dataframe with energy consumption/demand time series.
Column names: "time" and "eload".}

\item{temp_data}{A tidy dataframe with weather time series.
Column names: "time" and "temp".}

\item{data_is_quantity}{A Boolean. TRUE indicates interval data as quantity while
FALSE indicates interval data as rate.}
}
\value{
The orginal dataframe aggregated to hourly data interval.
}
\description{
Aggregate 15 minute dataframe to hourly interval data
}
