% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_eload_temp_df_to_monthly.R
\name{agg_eload_temp_df_to_monthly}
\alias{agg_eload_temp_df_to_monthly}
\title{Aggregate 15 minute, hourly or daily dataframe to a monthly interval.}
\usage{
agg_eload_temp_df_to_monthly(eload_data, temp_data, data_is_quantity = T,
  balancepoint_temp)
}
\arguments{
\item{eload_data}{A dataframe with energy consumption/demand time series.
Column names: "time" and "eload". If time series data interval is monthly,
add another column, 'days', listing the number of days in the corresponding
month.}

\item{temp_data}{A dataframe with weather time series.
Column names: "time" and "temp".}

\item{data_is_quantity}{A Boolean. TRUE indicates interval data as quantity while
FALSE indicates interval data as rate.}

\item{balancepoint_temp}{A numeric equal to the balancepoint temperature for
calculating HDD and CDD.}
}
\value{
The orginal dataframe aggregated to monthly data interval.
}
\description{
Aggregate 15 minute, hourly or daily dataframe to a monthly interval.
}
