% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_to_daily.R
\name{agg_to_daily}
\alias{agg_to_daily}
\title{Aggregate data to daily data}
\usage{
agg_to_daily(dataframe, operation)
}
\arguments{
\item{dataframe}{An nx2 time series dataframe.}

\item{operation}{A character string identifying the operation to perform while aggregating the data: Mean, Sum.}
}
\value{
The orginal dataframe aggregated to daily data interval
}
\description{
Aggregate data to daily data
}
