% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_to_hourly.R
\name{agg_to_hourly}
\alias{agg_to_hourly}
\title{Aggregate data to hourly data.}
\usage{
agg_to_hourly(dataframe, operation)
}
\arguments{
\item{dataframe}{An nx2 time series.}

\item{operation}{A character string identifying the operation to perform while aggregating the data: Mean, Sum.}
}
\value{
The original dataframe aggregated to hourly data interval.
}
\description{
Aggregate data to hourly data.
}
