% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate energy consumption and temperature data. Can be used to aggregate temperature data only as well.}
\usage{
aggregate(
  eload_data = NULL,
  temp_data = NULL,
  convert_to_data_interval = c("Hourly", "Daily", "Monthly"),
  temp_balancepoint = 65
)
}
\arguments{
\item{eload_data}{A dataframe with energy consumption time series. Column names: "time" and "eload". Allowed time intervals: less-than 60-mins, hourly, daily, monthly}

\item{temp_data}{A dataframe with weather time series. Column names: "time" and "temp". Allowed time intervals: less-than 60-mins, hourly, daily}

\item{convert_to_data_interval}{A character string indicating the time interval to which the dataframe should be aggregated: 'Hourly', 'Daily', and 'Monthly'}

\item{temp_balancepoint}{A numeric indicating the balancepoint for the temp_data dataframe}
}
\value{
a dataframe with energy consumption data and corresponding temperature data, aggregated to the indicated data interval. If energy consumption data is not available,
aggregated temperature data is returned.
}
\description{
\code{Aggregate energy consumption and temperature data to hourly, daily, or monthly data intervals.
This function uses dplyr for aggregation as opposed to xts. As a result, the timestamps are not shifted after aggregation.}
}
