% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_data.R
\name{align_data}
\alias{align_data}
\title{Align a timeseries to the start of a period.}
\usage{
align_data(data = NULL, original_alignment = NULL)
}
\arguments{
\item{data}{A dataframe with a timestamp and a variable. Column names: "time" and "data".}

\item{original_alignment}{A string indicating the original alignment of the input data. Options: 'start', 'end'.}
}
\value{
a dataframe with data aligned to the start of the period
}
\description{
\code{This function serves as a precursor to create_dataframe().}
}
