% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_avoided_energy_use.R
\name{calculate_avoided_energy_use}
\alias{calculate_avoided_energy_use}
\title{Calculate energy use avoided in the post-implementation period.}
\usage{
calculate_avoided_energy_use(modeled_data_obj)
}
\arguments{
\item{modeled_data_obj}{An object with modeled data results. For example, the output from 'model_with_TOWT' function.}
}
\value{
a list with the following components:
\describe{
  \item{savings_summary}{a dataframe with absolute savings value and the savings pct value.}
  \item{savings_df}{a dataframe with the savings achieved per time interval.}
  \item{pct_savings}{a numeric indicating the percent savings achieved.}
}
}
\description{
\code{This function computes the estimated avoided energy use using the predicted data of the post-implementation period.}
}
