% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_coverage.R
\name{calculate_coverage}
\alias{calculate_coverage}
\title{Calculate coverage of independent variables in the normalized conditions dataset.}
\usage{
calculate_coverage(site_temp_data, ref_temp_data,
  outlier_threshold = NULL, extrapolation_limit = NULL,
  site_temp_start = NULL, site_temp_end = NULL, data_interval = NULL)
}
\arguments{
\item{site_temp_data}{An nx2 dataframe with the temperature data of the facility. Colnames: time and temp.}

\item{ref_temp_data}{An nx2 dataframe with the hourly normalized temperature data corresponding to the facility's location. Colnames: time and temp.}

\item{outlier_threshold}{a numeric indicating the number of datapoints (hours or days, depending on the data interval) below which the temperature observation occurences will be considered an outlier.
Default values: 1 for daily data, up to 10 for hourly data.}

\item{extrapolation_limit}{A numeric indicating the percentage, beyond the minimum and maximum observed temperatures, up to which the data range may be extrapolated for model prediction.}

\item{site_temp_start}{a Date object indicating the baseline start date.}

\item{site_temp_end}{A Date object indicating the baseline end date.}

\item{data_interval}{Character string specifying the data time interval: "Hourly", "Daily, or "Monthly".}
}
\value{
a list with the following components:
\describe{
  \item{temp_coverage}{a data frame with temperature bins and associated frequencies of occurence from site temperature data and normalized temperature data.}
  \item{coverage_factor_summary}{a dataframe with calculated temperature and time coverage factors, counts of covered and uncovered datapoints in the normalized conditions dataset.}
  \item{temp_bin_summary}{a list with min, max, extrapolated min, and extrapolated max observed temperature values.}
}
}
\description{
\code{This function calculates the coverage of the independent variables in the normalized conditions dataset to
inform the user about the data range within which the model predictions are valid.}
}
