% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_savings_and_uncertainty.R
\name{calculate_norm_savings_and_uncertainty}
\alias{calculate_norm_savings_and_uncertainty}
\title{Calculate normalized savings and associated uncertainty as per ASHRAE Guideline 2014}
\usage{
calculate_norm_savings_and_uncertainty(
  baseline_model = NULL,
  baseline_stats = NULL,
  performance_model = NULL,
  performance_stats = NULL,
  normalized_weather = NULL,
  confidence_level = 90
)
}
\arguments{
\item{baseline_model}{Baseline model}

\item{baseline_stats}{Baseline model's corresponding statistics}

\item{performance_model}{Performance model}

\item{performance_stats}{Performance model's corresponding statistics}

\item{normalized_weather}{Weather dataset to be used for model prediction normalization}

\item{confidence_level}{Numeric corresponding to the confidence level to be used for savings uncertainty calculation}
}
\value{
a list with the following components:
\describe{
  \item{normalized_savings}{a dataframe with normalized baseline and performance predictions, and savings.}
  \item{normalized_savings_frac}{a numeric indicating the fraction of normalized savings against normalized baseline}
  \item{normalized_savings_unc}{a numeric indicating normalized savings uncertainty in energy units}
  \item{normalized_savings_unc_pct}{a numeric indicating normalized savings uncertainty in percent units}
  \item{confidence_leve}{a numeric indicating the confidence level at which the computation was performed.}
}
}
\description{
\code{This function calculates normalized (hourly or daily) savings achieved and the associated uncertainty.}
}
