% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_savings_uncertainty.R
\name{calculate_savings_uncertainty}
\alias{calculate_savings_uncertainty}
\title{Compute Savings Uncertainity.}
\usage{
calculate_savings_uncertainty(modeled_data_obj, savings_percent)
}
\arguments{
\item{modeled_data_obj}{An object with modeled data results. For example, the output from 'model_with_TOWT' function.}

\item{savings_percent}{a numeric indicating the observed/expected savings percentage}
}
\value{
a savings list object with savings uncertainty and minimum savings needed at 50\% uncertainty for 68\%, 90\%, 95\%, and 99.7\% confidence level.
}
\description{
\code{This function calculates the savings uncertainty, given the savings percent and a confidence level, as per ASHRAE Guideline 14.}
}
