% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_summary_statistics.R
\name{calculate_summary_statistics}
\alias{calculate_summary_statistics}
\title{Calculate model summary statistics. Applicable to training data only.}
\usage{
calculate_summary_statistics(modeled_data_obj = NULL)
}
\arguments{
\item{modeled_data_obj}{List with model results. Output from model_with_SLR, model_with_CP, model_with_HDD_CDD, and model_with_TOWT.}
}
\value{
a dataframe with five model statistics: "R_squared", "CVRMSE", "NDBE", "MBE", "#Parameters"
}
\description{
Calculate model summary statistics. Applicable to training data only.
}
