% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_temp_knots.R
\name{calculate_temp_knots}
\alias{calculate_temp_knots}
\title{Temperature knots calculation for model_with_TOWT}
\usage{
calculate_temp_knots(training_data = NULL, model_input_options = NULL)
}
\arguments{
\item{training_data}{Training dataframe and operating mode dataframe. Output from create_dataframe}

\item{model_input_options}{List with model inputs specified using assign_model_inputs}
}
\value{
a vector with calculated temperature knots
}
\description{
\code{This function determines temperature knots in the training dataset.}
}
