% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_temp_matrix.R
\name{create_temp_matrix}
\alias{create_temp_matrix}
\title{Create temperture bins for use in the TOWT algorithm.}
\usage{
create_temp_matrix(temp_col, temp_knots)
}
\arguments{
\item{temp_col}{A vector containing temperature data}

\item{temp_knot}{A vector specifying temperature knots to create temperature data segments}
}
\value{
A matrix with component temperatures
}
\description{
\code{This function generates temperature changepoints for subsequent use in energy data fitting.
This function, along with the function 'find_occ_unocc', is a utility function for another kWMV function: 'fit_TOWT_reg'.
This function is adapted from work by LBNL: \url{https://lbnl-eta.github.io/RMV2.0/}}
}
