% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_towt_weighted_reg.R
\name{create_towt_weighted_reg}
\alias{create_towt_weighted_reg}
\title{Create weighted linear regression models for energy use data using time-of-week and outside air temperature as independent variables.}
\usage{
create_towt_weighted_reg(time_col, eload_col, temp_col, pred_time_col,
  pred_temp_col, timescale_days = timescale_days,
  interval_minutes = interval_minutes,
  has_temp_knots_defined = has_temp_knots_defined,
  run_temperature_model = run_temperature_model,
  equal_temp_segment_points = equal_temp_segment_points,
  temp_segments = temp_segments, temp_knots_value = temp_knots_value,
  has_operating_modes = has_operating_modes,
  train_operating_mode_data = train_operating_mode_data,
  pred_operating_mode_data = pred_operating_mode_data,
  categories = categories)
}
\arguments{
\item{time_col}{time column of training data}

\item{eload_col}{eload column of training data}

\item{temp_col}{temp column of training data}

\item{pred_time_col}{time column of prediction data}

\item{pred_temp_col}{temp column of prediction data}

\item{timescale_days}{Numeric correspond to the timescale for weighting function.Default: NULL.
Change to improve accuracy of short term models.}

\item{interval_minutes}{Numeric for the interval period. Default: 15.}

\item{has_temp_knots_defined}{Boolean specifying whether the temp_knots are pre-defined or will be calculated by the algorithm}

\item{run_temperature_model}{Boolean specifying whether temperature should or should not be used in regression modeling.}

\item{equal_temp_segment_points}{Boolean specifying structure of temperature segments: equal number of points vs. equal segment length}

\item{temp_segments}{Numeric for number of temperature segments. Default: 6}

\item{temp_knots_value}{Vector specifying manually defined temperature knots.}

\item{has_operating_modes}{Boolean specifying whether the energy use profile has varying operating modes.}

\item{train_operating_mode_data}{dataframe with indicator variables for the various operating modes present in the model training period.}

\item{pred_operating_mode_data}{dataframe with indicator variables for the various operating modes present in the mdoel prediction period.}

\item{categories}{vector specifying names of the operating modes present.}
}
\value{
Weighted baseline and prediction matrices
}
\description{
\code{This function applies the hyperparamter, 'Weighting Factor', to the linear regressions generated by the function 'fit_towt_reg'.
It is a utility function for another kWMV function: 'model_with_towt'.
This function is adapted from work by LBNL: \url{https://lbnl-eta.github.io/RMV2.0/}}
}
