% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_demand_with_TOWT.R
\name{model_demand_with_TOWT}
\alias{model_demand_with_TOWT}
\title{Modeling electric demand Time-of-Week and Temperature.}
\usage{
model_demand_with_TOWT(training_data = NULL, prediction_data = NULL,
  model_input_options = NULL)
}
\arguments{
\item{training_data}{Training dataframe and operating mode dataframe. Output from create_dataframe}

\item{prediction_data}{Prediction dataframe and operating mode dataframe. Output from create_dataframe}

\item{model_input_options}{List with model inputs specified using assign_model_inputs}
}
\value{
a list with the following components:
\describe{
  \item{final_train_matrix} {training data matrix weighted according to timescale_days}
  \item{reg_out} {an lm object generated from fit_TOWT_reg}
  \item{final_pred_matrix} {prediction data matrix weighted according to timescale_days}
}
}
\description{
\code{This function builds an electric demand models using two algorithms: TOWT and MW.
This function is adapted from work by LBNL: \url{https://lbnl-eta.github.io/RMV2.0/}}
}
