% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_with_TOWT.R
\name{model_with_TOWT}
\alias{model_with_TOWT}
\title{Generate an energy data model using the Time-of-Week and Temperature algorithm.}
\usage{
model_with_TOWT(training_data = NULL, prediction_data = NULL,
  timescale_days = NULL, interval_minutes = 15,
  has_temp_knots_defined = FALSE, run_temperature_model = TRUE,
  equal_temp_segment_points = T, temp_segments_numeric = 6,
  temp_knots_value = c(40, 45, 50, 60, 65, 90),
  has_operating_modes = FALSE, train_operating_mode_data = NULL,
  pred_operating_mode_data = NULL, data_interval = NULL,
  data_units = NULL)
}
\arguments{
\item{training_data}{Training period dataframe, where the columns correspond to the time steps (time), the energy load (eload), and the temperature (Temp).}

\item{prediction_data}{Prediction period dataframe, where the columns correspond to the time steps (time), the energy load (eload), and the temperature (Temp).}

\item{timescale_days}{Numeric correspond to the timescale for weighting function.Default: NULL.
Change to improve accuracy of short term models.}

\item{interval_minutes}{Numeric for the interval period. Default: 15.}

\item{has_temp_knots_defined}{Boolean specifying whether the temp_knots are pre-defined or will be calculated by the algorithm}

\item{run_temperature_model}{Boolean specifying whether temperature should or should not be used in regression modeling.}

\item{equal_temp_segment_points}{Boolean specifying structure of temperature segments: equal number of points vs. equal segment length}

\item{temp_segments_numeric}{Numeric for number of temperature segments. Default: 6}

\item{temp_knots_value}{Vector specifying manually defined temperature knots.}

\item{has_operating_modes}{Boolean specifying whether the energy use profile has varying operating modes.}

\item{train_operating_mode_data}{dataframe with indicator variables for the various operating modes present in the model training period.}

\item{pred_operating_mode_data}{dataframe with indicator variables for the various operating modes present in the mdoel prediction period.}

\item{data_interval}{Character string specifying the data time interval: "Hourly", "Daily, or "Monthly".}

\item{data_units}{energy data's units}
}
\value{
a TOWT_baseline object, which is a list with the following components:
\describe{
  \item{TOWT_model}{an object that has been created by the function create_TOWT_weighted_reg,
   and which correspond to the TOWT model.}
  \item{training_data}{a dataframe corresponding to the training data after the
  cleaning and filtering function were applied, fitted values, and residuls}
  \item{goodness_of_fit}{a data frame that contains the goodness of fitting metrics.}
  \item{rsdl}{a data frame containing all residual values}
  \item{normality metrics}{a list with details on residuals' skewness and kurtosis.}
  \item{energy use summary}{Summed baseline, post-implementation, and adjusted baseline energy use values. Assumes training dataset is the
  energy project's baseline energy dataset.}
  \item{post_implementation_data}{a dataframe corresponding to the post-implementation dataset along with predicted values.}
}
}
\description{
\code{This function builds an energy use model using two algorithms: TOWT and MW.
This function is adapted from work by LBNL: \url{https://lbnl-eta.github.io/RMV2.0/}}
}
