% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{school_eload}
\alias{school_eload}
\title{Energy Use data of a school.}
\format{A data frame with 8760 rows and 2 variables:
\describe{
  \item{time}{timestamp, format mm/dd/yy hh:mm:ss}
  \item{eload}{energy consumption data aggregated to the hourly interval, in kWh}
  ...
}}
\usage{
school_eload
}
\description{
A dataset containing the hourly average temperature for the time period: 01/01/2018 - 12/31/2018.
}
\keyword{datasets}
