% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{school_op_mode}
\alias{school_op_mode}
\title{Operating mode information data of a school.}
\format{A data frame with 8760 rows and 5 variables:
\describe{
  \item{time}{timestamp, format mm/dd/yy hh:mm:ss}
  \item{school_holidays}{indicator variable marking the school holidays in the given time period}
  \item{summer maintenance}{indicator variable marking periods of maintenance during the given time period}
  \item{summer school}{indicator variable marking the days of summer school in the given time period}
  \item{pre-class ramp up}{indicator variable marking days of pre-class activity ramp up before the Fall session}
  ...
}}
\usage{
school_op_mode
}
\description{
A dataset containing the hourly average temperature for the time period: 01/01/2018 - 12/31/2018.
}
\keyword{datasets}
