% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{school_temp}
\alias{school_temp}
\title{Temperature data for a school.}
\format{A data frame with 8760 rows and 2 variables:
\describe{
  \item{time}{timestamp, format mm/dd/yy hh:mm:ss}
  \item{temp}{outside air temperature corresponding to the timestamp, in deg F}
  ...
}}
\usage{
school_temp
}
\description{
A dataset containing the hourly average temperature for the time period: 01/01/2018 - 12/31/2018.
}
\keyword{datasets}
