% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temp}
\alias{temp}
\title{Temperature data.}
\format{
A data frame with 732 rows and 2 variables:
\describe{
  \item{time}{timestamp, format mm/dd/yy}
  \item{temp}{outside air temperature corresponding to the timestamp, in deg F}
  ...
}
}
\usage{
temp
}
\description{
A dataset containing the daily average temperature for the time period: 04/07/2017 - 04/07/2019.
}
\keyword{datasets}
