% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TMY3}
\alias{TMY3}
\title{TMY3 data.}
\format{
A data frame with 8760 rows and 2 variables:
\describe{
  \item{time}{timestamp, format mm/dd/yy hh::mm::ss}
  \item{temp}{outside air temperature corresponding to the timestamp, in deg F}
  ...
}
}
\usage{
TMY3
}
\description{
A dataset containing the TMY3 dataset for demonstrating savings normalization .
}
\keyword{datasets}
