% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_coverage.R
\name{calculate_coverage}
\alias{calculate_coverage}
\title{Calculate coverage of independent variables in the normalized conditions dataset.}
\usage{
calculate_coverage(dataframe = NULL, ref_temp_data = NULL,
  outlier_threshold = NULL, extrapolation_limit = NULL)
}
\arguments{
\item{dataframe}{Training or Prediction dataframe from create_dataframe}

\item{ref_temp_data}{An nx2 dataframe with the hourly normalized temperature data corresponding to the facility's location. Colnames: time and temp.}

\item{outlier_threshold}{a numeric indicating the number of datapoints (hours or days, depending on the data interval) below which the temperature observation occurences will be considered an outlier.
Default values: 1 for daily data, up to 10 for hourly data.}

\item{extrapolation_limit}{A numeric, beyond the minimum and maximum observed temperatures, up to which the data range may be extrapolated for model prediction. Default: 0.05}
}
\value{
a list with the following components:
\describe{
  \item{temp_coverage}{a data frame with temperature bins and associated frequencies of occurence from site temperature data and normalized temperature data.}
  \item{coverage_factor_summary}{a dataframe with calculated temperature and time coverage factors, counts of covered and uncovered datapoints in the normalized conditions dataset.}
  \item{temp_bin_summary}{a list with min, max, extrapolated min, and extrapolated max observed temperature values.}
}
}
\description{
\code{This function calculates the coverage of the independent variables in the normalized conditions dataset to
inform the user about the data range within which the model predictions are valid.}
}
