% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_TOWT_model_predictions.R
\name{calculate_TOWT_model_predictions}
\alias{calculate_TOWT_model_predictions}
\title{Calculate TOWT model predictions}
\usage{
calculate_TOWT_model_predictions(training_data = NULL,
  prediction_data = NULL, modeled_object = NULL)
}
\arguments{
\item{training_data}{Training dataframe and operating mode dataframe. Output from create_dataframe}

\item{prediction_data}{Prediction dataframe and operating mode dataframe. Output from create_dataframe}

\item{model_input_options}{List with model inputs specified using assign_model_inputs}
}
\value{
a list with the following components:
\describe{
  \item{predictions} {dataframe with model predictions}
}
}
\description{
\code{This function calculates predictions from model_with_TOWT}
}
