% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataframe.R
\name{create_dataframe}
\alias{create_dataframe}
\title{Generate training or prediction dataframes.}
\usage{
create_dataframe(
  eload_data = NULL,
  temp_data = NULL,
  operating_mode_data = NULL,
  additional_independent_variables = NULL,
  additional_variable_aggregation = c(sum, median),
  start_date = NULL,
  end_date = NULL,
  convert_to_data_interval = c("15-min", "Hourly", "Daily", "Monthly"),
  temp_balancepoint = 65
)
}
\arguments{
\item{eload_data}{A dataframe with energy consumption time series. This dataframe should only be energy consumption data and not demand data. Column names: "time" and "eload". Allowed time intervals: 15-min, hourly, daily, monthly. The 'time' column must have Date-Time object values.}

\item{temp_data}{A dataframe with weather time series. Column names: "time" and "temp". Allowed time intervals: 15-min, hourly, daily, monthly. The 'time' column must have Date-Time object values.}

\item{additional_independent_variables}{An optional dataframe for adding independent variables to the regression. This argument is a replacement for the older 'operating_mode_data' argument.}

\item{additional_variable_aggregation}{A vector with aggregation functions for each of the variables in 'additional_independent_variables'.
Usage example: c(sum, median) implies two additional independent variables. The first variable will be summed over the specified data interval
and the median of the second variable will be taken over the specified data interval. Permissible aggregation functions: sum, mean, median}

\item{start_date}{A character string, of the format "mm/dd/yyy hh:mm", indictating start date and time of the intended dataframe}

\item{end_date}{A character string, of the format "mm/dd/yyy hh:mm", indictating end date and time of the intended dataframe}

\item{convert_to_data_interval}{A character string indicating the time interval to which the dataframe should be aggregated: '15-min', 'Hourly', 'Daily', and 'Monthly'}

\item{temp_balancepoint}{A numeric indicating the balancepoint for the temp_data dataframe}
}
\value{
a dataframe with energy consumption data, temperature data and additional variable data
}
\description{
\code{This function creates a dataframe, combining eload, temp, and additional variable data.
 It assumes that the input data is aligned to the start of a time period and outputs a dataframe aligned to the end of time periods.
 NA values are ignored during aggregation.}
}
