% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataframe.R
\name{create_dataframe}
\alias{create_dataframe}
\title{Generate training or prediction dataframe.}
\usage{
create_dataframe(eload_data = NULL, temp_data = NULL,
  operating_mode_data = NULL, start_date = NULL, end_date = NULL,
  convert_to_data_interval = c("Hourly", "Daily", "Monthly"),
  temp_balancepoint = 65)
}
\arguments{
\item{eload_data}{A dataframe with energy/demand consumption time series. Column names: "time" and "eload". Allowed time intervals: less-than 60-mins, hourly, daily, monthly}

\item{temp_data}{A dataframe with weather time series. Column names: "time" and "temp". Allowed time intervals: less-than 60-mins, hourly, daily}

\item{operating_mode_data}{A dataframe with indicator variables for different energy use profiles withtin the stated time period.
A time column, corresponding to eload_data and/or temp_data must be included.}

\item{start_date}{A character string, of the format "mm/dd/yyy hh:mm", indictating start date and time of the intended dataframe}

\item{end_date}{A character string, of the format "mm/dd/yyy hh:mm", indictating end date and time of the intended dataframe}

\item{convert_to_data_interval}{A character string indicating the time interval to which the dataframe should be aggregated: 'Hourly', 'Daily', and 'Monthly'}

\item{temp_balancepoint}{A numeric indicating the balancepoint for the temp_data dataframe}
}
\value{
a list with energy consumption data and corresponding temperature data, aggregated to the indicated data interval
}
\description{
Generate training or prediction dataframe.
}
