% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_temp_matrix.R
\name{create_temp_matrix}
\alias{create_temp_matrix}
\title{Temperature matrix for model_with_TOWT.}
\usage{
create_temp_matrix(temp_col = NULL, temp_knots = NULL)
}
\arguments{
\item{temp_col}{A vector containing temperature data}

\item{temp_knots}{A vector specifying temperature knots to create temperature data segments}
}
\value{
A matrix with component temperatures
}
\description{
\code{This function generates temperature changepoints for subsequent use in energy data fitting}
}
