% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_occ_unocc.R
\name{find_occ_unocc}
\alias{find_occ_unocc}
\title{Divide time-of-energy-use into occupied and unoccupied modes for use TOWT algorithm.}
\usage{
find_occ_unocc(interval_of_week, eload_col, temp_col,
  interval_minutes = 15)
}
\arguments{
\item{interval_of_week}{Unique indicator for each 15-min interval of the week.}

\item{eload_col}{eload column of training data.}

\item{temp_col}{temp column of training data.}

\item{interval_minutes}{Numeric of length of a Time Of Week interval as input variables. Deafult: 15.}
}
\value{
A matrix with unique times of week along with a 0 or 1, representing occupied and unoccupied modes.
}
\description{
\code{This function generates occupancy schedules for subsequent use in energy data fitting.
This function, along with the function 'create_temp_matrix', is a utility function for another kWMV function: 'fit_towt_reg'.
This function is adapted from work by LBNL: \url{https://lbnl-eta.github.io/RMV2.0/}}
}
