% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_occ_unocc.R
\name{find_occ_unocc}
\alias{find_occ_unocc}
\title{Occupancy detection algorithm for model_with_TOWT.}
\usage{
find_occ_unocc(training_data = NULL, model_input_options = NULL)
}
\arguments{
\item{data}{dataframe and operating mode dataframe. Output from create_dataframe}
}
\value{
A matrix with unique times of week along with a 0 or 1, representing occupied and unoccupied modes.
}
\description{
\code{This function generates occupancy schedules for subsequent use in energy data fitting.}
}
